+++ orderedArray.h
@@ -23,7 +23,7 @@ public:
     void insertUnique(const string& word, int occurrences);    //Get a word and occurrence number and place it in the correct spot.
     int search(const string& word);                                  //Search for a word and return it's occurrences (0 for none).
     void remove(const string& word);                                 //Remove word from the array and readjust the order.
-    int getSize() const;                                                  //Returns the size of the array.
+    long getSize() const;                                                  //Returns the size of the array.
